pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smokeshow"
version = "0.1.0"
description = "OpenTelemetry instrumentation for Playwright E2E tests"
requires-python = ">=3.9"
dependencies = [
    "playwright>=1.40.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "pytest-cov>=4.1"]

[tool.hatch.build.targets.wheel]
packages = ["src/smokeshow"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
